Player physics tweaks
+ The player moves more slowly when crouching
* The player has less air control
* The player now jumps when crouching on a non-one-way platform
+ Pawn.airControl - 25% by default
* Pawn.moveH() no longer takes input; it instead takes a max speed multiplier and an acceleration multiplier

+++ js/GameObjects/Pawn/PawnStates.js
@@ -46,7 +46,7 @@ const PawnStates={
             if(pawn.input.crouch) pawn.state=PawnStates.crouching;
             if(!pawn.isGrounded) pawn.state=new PawnStates.inAir();
 
-            pawn.moveH(pawn.input.move);
+            pawn.moveH();
             pawn.moveV();
         }
     },
@@ -80,7 +80,7 @@ const PawnStates={
         };
         this.update=(pawn)=>{
             if(!PawnStates.isPawn(pawn)) return;
-            pawn.moveH(pawn.input.move);
+            pawn.moveH(1,pawn.airControl);
             pawn.moveV();
             if(pawn.isGrounded)pawn.state=PawnStates.idle;
             if(pawn.input.onJump)pawn.launch({y:-375},true); // set state to  jumping
@@ -95,7 +95,7 @@ const PawnStates={
         },
         update(pawn){
             if(!PawnStates.isPawn(pawn)) return;
-            pawn.moveH(pawn.input.move);
+            pawn.moveH(1,pawn.airControl);
             pawn.moveV(.4); // less gravity when jumping
             if(pawn.vy>0 || pawn.input.jump==false) pawn.state=new PawnStates.inAir();
 
@@ -109,15 +109,17 @@ const PawnStates={
         update(pawn){
             if(!PawnStates.isPawn(pawn)) return;
             
-            pawn.moveH(pawn.input.move);
-            pawn.moveV();
 
-            if(pawn.input.onJump) {
-                pawn.drop();
-                return;
-            }
             if(!pawn.input.crouch) pawn.state = PawnStates.idle;
             if(!pawn.isGrounded) pawn.state=new PawnStates.inAir();
+            if(pawn.input.onJump) {
+                if(pawn.onOneway)pawn.drop();
+                else {
+                    pawn.launch({y:-375},true); // set state to  jumping
+                }
+            }
+            pawn.moveH(.5);
+            pawn.moveV();
         }
     },
     sneaking:{
@@ -141,7 +143,7 @@ const PawnStates={
             if(!pawn.isGrounded)pawn.state=new PawnStates.inAir();
             if(pawn.input.onJump)pawn.launch({y:-375},true);
             if(pawn.input.crouch)pawn.state=PawnStates.crouching;
-            pawn.moveH(pawn.input.move);
+            pawn.moveH();
             pawn.moveV();
         }
     },
